+++ front/src/modules/object-record/field/meta-types/display/components/CurrencyFieldDisplay.tsx
@@ -1,4 +1,4 @@
-import { CurrencyDisplay } from '@/ui/field/components/CurrencyDisplay';
+import { CurrencyDisplay } from '@/ui/field/display/components/CurrencyDisplay';
 
 import { useCurrencyField } from '../../hooks/useCurrencyField';
