+++ front/src/modules/object-record/field/meta-types/display/components/EmailFieldDisplay.tsx
@@ -1,4 +1,4 @@
-import { EmailDisplay } from '@/ui/field/components/EmailDisplay';
+import { EmailDisplay } from '@/ui/field/display/components/EmailDisplay';
 
 import { useEmailField } from '../../hooks/useEmailField';
