+++ front/src/modules/object-record/field/meta-types/display/components/LinkFieldDisplay.tsx
@@ -1,4 +1,4 @@
-import { LinkDisplay } from '@/ui/field/components/LinkDisplay';
+import { LinkDisplay } from '@/ui/field/display/components/LinkDisplay';
 
 import { useLinkField } from '../../hooks/useLinkField';
