+++ front/src/modules/object-record/field/meta-types/display/components/UuidFieldDisplay.tsx
@@ -1,5 +1,5 @@
 import { useUuidField } from '@/object-record/field/meta-types/hooks/useUuidField';
-import { TextDisplay } from '@/ui/field/components/TextDisplay';
+import { TextDisplay } from '@/ui/field/display/components/TextDisplay';
 
 export const UuidFieldDisplay = () => {
   const { fieldValue } = useUuidField();
