+++ front/src/modules/object-record/field/meta-types/input/components/LinkFieldInput.tsx
@@ -1,8 +1,8 @@
-import { TextInput } from '@/object-record/field/meta-types/input/components/internal/TextInput';
+import { TextInput } from '@/ui/field/input/components/TextInput';
 
+import { FieldInputOverlay } from '../../../../../ui/field/input/components/FieldInputOverlay';
 import { useLinkField } from '../../hooks/useLinkField';
 
-import { FieldInputOverlay } from './internal/FieldInputOverlay';
 import { FieldInputEvent } from './DateFieldInput';
 
 export type LinkFieldInputProps = {
