+++ front/src/modules/ui/field/display/components/CurrencyDisplay.tsx
