+++ front/src/modules/ui/field/display/components/DoubleTextDisplay.tsx
