+++ front/src/modules/ui/field/display/components/EllipsisDisplay.tsx
