+++ front/src/modules/ui/field/display/components/EmailDisplay.tsx
