+++ front/src/modules/ui/field/display/components/LinkDisplay.tsx
