+++ front/src/modules/ui/field/display/components/MoneyDisplay.tsx
