+++ front/src/modules/ui/field/display/components/NumberDisplay.tsx
