+++ front/src/modules/ui/field/display/components/PhoneDisplay.tsx
