+++ front/src/modules/ui/field/display/components/TextDisplay.tsx
