+++ front/src/modules/ui/field/display/components/__stories__/PhoneInputDisplay.stories.tsx
