+++ front/src/modules/ui/field/input/components/RatingInput.tsx
